+++ src/compiler/file-ctx.ts
@@ -5,6 +5,9 @@
  * @since 0.0.3
  */
 
+import { CompilationUnitContext } from "./parser";
+import { KipperParseStream } from "./parse-stream";
+
 /**
  * The file context class used to represent a file in a compilation.
  *
@@ -13,5 +16,36 @@
  * @since 0.0.3
  */
 export class KipperFileContext {
+  /**
+   * The private '_stream' that actually contains the instance,
+   * which is used inside the getter 'stream'
+   * @private
+   */
+  private readonly _stream: KipperParseStream;
 
+  /**
+   * The private '_startItem' that actually contains the instance,
+   * which is used inside the getter 'startItem'
+   * @private
+   */
+  private readonly _startItem: CompilationUnitContext;
+
+  constructor(stream: KipperParseStream, startItem: CompilationUnitContext) {
+    this._stream = stream;
+    this._startItem = startItem;
+  }
+
+  /**
+   * Returns the {@link KipperParseStream} which contains the raw file data.
+   */
+  get stream() {
+    return this._stream;
+  }
+
+  /**
+   * Returns the start item of the parser tree (top item).
+   */
+  get startItem() {
+    return this._startItem;
+  }
 }