+++ src/kipper-standalone.ts
@@ -1,12 +1,11 @@
-import { KipperCompiler } from "./compiler";
-import { KipperLogger, LogLevel, LogLevelNames } from "./logger";
+import * as kipper from ".";
 
-function emitHandler(level: LogLevel, msg: string): void {
-  console.log(`[${LogLevelNames.get(level)}] ${msg}`)
+function emitHandler(level: kipper.LogLevel, msg: string): void {
+  console.log(`[${kipper.LogLevelNames.get(level)}] ${msg}`)
 }
 
-let logger: KipperLogger = new KipperLogger(emitHandler);
-let compiler: KipperCompiler = new KipperCompiler(logger);
+let logger: kipper.KipperLogger = new kipper.KipperLogger(emitHandler);
+let compiler: kipper.KipperCompiler = new kipper.KipperCompiler(logger);
 
 // Test message
 logger.info("Testing functionality");